fix: use textfield instead of charfield in schema
CharField was initially choosen with the assumptions, that it
is more performance and/or space efficient. That is not true.
In postgres varchar and text are stored alike and its safer
here to use text, because some fields  might contain long values
and break the input if they don't fit the max_length constraint.