refactor: use only django orm and transactions
We were using pandas.to_sql and django for inserting data into the
the database. They behave differently with datetimes that have
timezones. Pandas.to_sql does not support timezones. This uses only
the django ORM for both inserts and updates.

This also wraps the updates in a transaction to improve the
performance.