feat: serve webpack assets over https
We do this, so that we can proxy the dev server over http. That
makes it possible to run the mobile app over https and connect from
there to the server. If the server is on http, the mobile app would
error because of cross-origin.

DOCKER_HOST is relic we don't need anymore and is removed in the
process.