SLEEP-320 Patient API filters
The subqueries are pretty efficient even if they generate nasty and lengthy SQL. However, the default date range for tests is taking a lot of time because it forces a join patient-case-test just to get the dates. The lookup in cases_case and patient_test are pretty efficient because they use indexes but the patient_patient table is fully scanned to retrieve the results.