SLEEP-426 Improve performance on patient download
The biggest gain comes from the change of queryset_iterator implementation and its use in the iter_items function. The CSV export of patients goes from 200KB/min to 250Kb/sec. (mind that /min becomes /sec !) It also removes the constraint of sorting by primary key.
The other improvements are only mild improvements.
Exporting all of the database in the patients tab is still going to take forever but exporting the tests in CSV works just fine for over a million records. Patients XLSX is also slower than CSV since it's exporting several sheets. So don't try to export the whole database.