SLEEP-526 Allow village search on its aliases too
This looks trivial but it's not. You can't do an `__icontains` on an array, it'd search wether the array contains the provided string exactly. The key is to use `unnest` that could be wrapped into a `Func(F('aliasas'), function='unnest')` but it's a special type that will actually return several rows, that cannot be used in where expressions and have a weird join behavior that's called a lateral join. I tried many ways. The sql `where exists (select it from unnest(gv.aliases) it where it ilike '%avkiv%')` works fine but Django won't take it because it's a special type and we cannot do `Exists(RawSQL(...)`
So the only way I found was to annotate with a count of matching aliases. I am not too sure about the RawSQL with a user-provided parameter but it resisted my SQL injection attempts so ¯\_(ツ)_/¯

The alternative was to concat the array to text and use icointains on that. I'm sure this would have been a disaster waiting to happen.