SLEEP-541 Adding latest_test_date to cases API ⤵️
This change actually brings a bigger change in the code. So far, we were using CaseView to query the records. CaseView inherits from CaseAbstract, just as Case does. When we render the as_dict of CaseView, we do a Case.as_dict() to render the Case part of it. The problem is that test_set is not available in CaseView and is necessary to have any kind of reference to the tests and get their latest date.
So I added a couple of methods to CaseView to actually "annotate" a Case query to mimic the CaseView fields. Not all of them are available right now but I think that we should avoid the composite fields on all test results and rather use the patient_test table.

At some point, we'll get a request to sort on that field and that will require additional processing to avoid a nasty query like adding a denormalized "latest_test_date" into the cases_case table.