SLEEP-546 Fix the user/profile delete to disable the user + fix tester_id
The profile.delete() operation was in fact deleting only the profile and cascading to the patient_test.tester_id.
This commit uses the is_active flag, allows to create a user that already exists to re-enable it and restore the tester_id from the devicedb that still points to the user.