SLEEP-563 Fix inverted name/postname in some cases
This was a tricky one:
The mapping.py had mixed up the mapping for post name and last name between historic/pv and mobile.
Add the fact that the name in cases_case was confusing (name, lastname) and the fact that creating the patient straightened the data, you get a weird mix-up.

This update renames the cases_case "name" to "postname" and switches the appropriate names.

This will obviously require a refresh of the patient duplicates that is going to take a while.