Fix log format issue, add formating to log file
logger in iaso and beanstalk_worker were not using the correct format
(not printing date and module name) nor did they use the log level.

We now use this format for all modules. and correcty use the log level
env var.

Logging file in production didn't print the date (or use the formatting),
this is now corrected.

Note that at the moment there is no logger under hat