IA-994 Speed up the group filter selector
Prefetch related queries
and only return necessary data. Previously we were returning the DataSource and all the versions with no relation to the group

This diminish the response size, the returned json is now 36k from
216.25kb before.
And the number of SQL queries on my local database goes from 1300 to 7.

Request Reponse time have been improved from 4622 ms (4 seconds) to
831ms.