POLIO-312 Optimize the vaccine stock endpoint
this endpoint is one of several used to generate the Vaccine dashboard in PowerBI
But PowerBi is timing out on it, causing the whole refresh of the
dashboard to fail. So this PR optimise it to return faster.

This code optimize the DB requests and cache the campaign finding code.
So generation is down from one minute to 1s on my computer for a copy of
the prod database.
Number of DB request goes from 7268 query to 160 queries (divided by 45!)

Note for further possible improvement
1. the file is still arround 9mb so the transfer time may stay high. We
 could drop data unused in the dashboard to improve that.
2. timing this is with the url cache,  so the first request will still be
slower than the subsequent one. We could eventually do cache priming before powerbi
hit the service

We will of course need to check in prod that this improvement is
sufficient for PowerBI but is a usefull improvement nonetheless