Explicitly require the backend to be set
The BACKGROUND_TASK_SERVICE environment variable will now decide which
background workers task service to use.

Since the webapp needs to be able to enqueue jobs we couldn't just
rely on the WORKER flag and the existnence of a BEANSTALK_SQS_URL,
that's why I decided to be more explicit about it.

And break if the wrong task service is defined.