Support setting a `STATIC_URL` env variable as CDN
What this means is that we're using S3 to store uploaded files, but
we're using the CDN to serve the (already built) static assets.

If you don't supply a `STATIC_URL` it is assumed that if you run with
`USE_S3` that the assets will also be found in that bucket.