POLIO-691 Speed up calendar map (campaign shapes) :fire:
It was the shapes endpoint that returned the geojson that got very slow.
We speed it up From 11 minutes to 9seconds by now used the
simplified_geom instead of the geom.

This code is only a port of the fix, we also need to manually run the following command manually
to recalculate the simplified_geom on iaso_orgunit for the newly added
org unit since the last time it was ran.

update iaso_orgunit set simplified_geom = ST_Simplify(geom::geometry, 0.01, true);

We need a more permanent fix in the futures to keep them up to
date, but this is outside the scope of this Hotfix PR.