POLIO-780 Crash when encoding campaign scope
Happened with the Botswana campaign
Attempt to FIX POLIO-780.
I made the change from the error on production since I can't reproduce it locally and can't manage to reproduce consistently in production
I assume this is a concurrency error following the order in which the server respond to the request.
Nevertheless this check was necessary.