IA-1326 Fix issue with EMPTY GEOM
we had OrgUnit with empty geom in the database that could not be edited anymore because they had the issue
`Column has Z dimension but geometry does not`

they were created via the GPKG import or directly in the database by inserting a POINT Z EMPTY,
but they couldn't be edited anymore from django because we got a 2d
empty point in return.

After discussion with the geo type we decided that it's better to
replace empty point with `null` value in database.

So this patch fix the issue by:
replacing Empty geom with None during GPKG import
Allow editing existing org unit by fixing their value in the save()