IA-2232 Fix duplicate orgunit in registry missing submissions
This is an error in the completeness stats endpoint, I have found out that the trigger  is using : `order=orgunit__name`
Which do a supplementary outer join . which duplicate row and produce inconsistent count result. This mean there are a lot of rows but only the first two are taken.

For the moment I found a fix by changing the paramter to order=name in the frontend

But they are probably other parameter that will cause this problem, so we will probably need a better fix in the long term.
  I checked `order by parent__name` don’t trigger it.

I tried to fix this using a distinct but this cause other issues, and
don't prevent the join

This error is reproducible directly from the stock data of seed_test_data.

To reproduce the error
-
Do the seed data import
Go to registry
Sierra Leone -> Western Area -> Event Tracker
Missing Submission(s)

What happen:

Rural Western Area is listed twice

What should happen:

It should list: Freetown and Rural Western Area.