Adding Locust to stress test Iaso
This contains a `locustfile.py` with a first try at simulating a mobile
flow. The Locust users will synchronize sometimes and upload often.
A docker configuration has been added to easily launch Locust along with
a small README.

The file `instance.xml` still needs to be filled and the ids corrected.

IA-2129