IA-2259 Make a new version for orgunit type api
Confusingly the V2 is the current behaviour with the new allow_creating_sub_unit_types field (separate from sub_unit_types)
the existing url is with the new behaviour where we map sub_unit_types to allow_creating_sub_unit_types.
so the mobile application don't need any update to alter the url.

At some point we shoud restrict the action on the current url to only
what the mobile app support, to reduce the amount of code needed for
maintenance