TRIA-7 Rename hat.vector_control App to hat.api_import
This Django app was still named like the old one and giving lot's of
name clashes with the Trypelim plugin.
I figured a rename was a good idea anyways, since the old name did not
reflect at all it's function.
I also did a cleanup in the migrations, only leaving the squashed one.