POLIO-1477: add DELETE to vaccin_stock endpoint
- Add DELETE
- Add test

This is a hard delete, like in supplychain. It was provided as a way to enable correction of bad input during manual input of historical data. If a soft delete is required, it should be done in a separate PR, and include supplychain, which has a hard delete as well