IA-2967 Add calculated fields to submissions filter
In the dashboard, on the form submissions page, when you filter results
on a specific form, you can choose to filter on attributes in the forms
(modal with title "Search in submitted fields").

Here, I've now made the calculated fields also available since it makes
a lot of sense to search on these in the light of searching in reference
forms for entities.