IA-2968 Add link to entity (if any) on form details page
Some improvements on the visual linking of entities with form
submissions.
If a form submission/instance has a `entity_id`, we display this on the
submission details page.
With this in place, there's no reason why we should show a simplified
page of a form submission when navigating to it via the entities (e.g.
images are not displayed).
Instead we can simply link directly to the form.