SLEEP-1337 Improvement: Create APIImport record...
as a first step in the view.
This way we keep an easier trace of failed calls.

Note: I did not use the @safe_api_import decorator, because I want to
pass the api_import instance along to the worker. Since in the decorator
it is created "around" the view method, and since it is wrapped in a db
transaction, this can give weird behaviour with the record not yet
existing in the db.