SLEEP-1448 Bulk upload: support form update and bugfix
- Iterate over the instances.json file instead of the folders to avoid
files not being found in case of a form that already exists. In this
case the instance already exists in the db, but potentially with a
different file_name if it was uploaded from a different tablet (nearby
share does not preserve filenames).
- If during processing we get a form that already exists, an for which the
bulk upload contains a newer form (with a more recent `updated_at`
timestamp), we update the form and set the new `source_updated_at`.
- Ignore .xml files when processing attachments to circumvent a bug in
the mobile app