IA-3253 Fix: Entity.MultipleObjectsReturned on /api/entities/{pk}/
When a user has geo-restrictions, the filter_for_user() will add an
INNER JOIN to the query, resulting in duplicates returned.

Adding the retrieve action explictely and adding a `distinct` to the
query fixes the issue.

Also: no need to add the full instance model in the serializer, it adds
a lot of queries and weight to the call and it's not used in the
dashboard. By simply changing to `as_dict()`, the payload is only 1/5th
of what it was before.