SLEEP-1459 Add retries with exponential backoff...
to XML file downloads.

Implemented because of the bulk upload.
The bulk upload downloads a lot of files rapidly from S3 (the XML files)
This can result in error:
HTTP Error 503: Slow Down

To handle this we add a 3 retries by default. With a 1 sec, 2 sec and 4
sec delay.
For the bulk uploads we can up this to 8 retries (the last one will have
a 128 sec delay).