SLEEP-1439 Improve Task permission to enable loading states...
such as for the generation of the mobile app setup zip.

Basically: keep the tasks endpoint permissions as is, but allow a user
to query the state of a task launched by themselves, even if they don't
have the "DATA_TASKS" permission.

E.g. For the download of the mobile app setup zip, we display a progress
bar. To display this, we fetch the task that was just created. But if
the user doesn't have the DATA_TASKS permission, this doesn't work.
It makes sense to allow this, but only on the `retrieve` endpoint and
only for tasks created by the user themself.