Fix `created_at` is `null` when downloading OUCRC
When creating on-the-fly configuration to restrict OrgUnit Change Request creation,
the `created_at` and `updated_at` dates were not set. This resulted in the backend
sending `null` for those required field.

Those fields are now set to the current date and time to fix the issue in the mobile app.

IA-3878