Add a `Type` to OUCRC
OUCRC now have a type to differentiate them between OrgUnit edition and creation.
The UI has been adapted to take this into consideration and the APIs as well.

In the mobile endpoint, a new `include_creation` parameter has been added not to break
with the existing applications which do not expect two kinds of OUCRC.

IA-4004, IA-4005, IA-4006