Refactoring: Simplify MetricValue filtering query
Based on the correct final SQL query I wondered if the `SUM(CASE ...`
annotation could be somehow converted to `EXISTS`.
Based on this, Claude Code showed me we could do the same with
subqueries in regular `WHERE` clauses. The result is less generic, but I
think easier to wrap your head around.