Fixing the n+1 query problem we had on csv export for the property is_reference_instance
It did not appear in the query count because it's a streaming API, which explains why we missed it.
I was not able to fix this using a prefetch, so I took another approach that is similar in terms of logic (prefetching) but using a separate query
The value of the field was tested in an existing test, I didn't have to add one.