Fix imports fails due to storage logs
`process_mobile_bulk_upload` was wrapped into one giant transaction.
This was problematic because we can have multiple zips each with the **full** storage logs.
The storage logs could point on entities that are created in one of the zips.
If we have entities created in at least two zips, the storage logs can never succeed and therefore
the import of all the zips is blocked.

Also, I extracted the logic of finding an entity to attach the storage logs to the proper entity even
when there are duplicates.

WC2-852