Fix looking up instances using JSONLogic with suffixed questions
Due to the mobile application's need to assign a type to some calculated questions, the
JSONField `Instance.json` might contain keys with a suffix like this: `__int__`.
This collides with DRF's way to declare accessing sub-objects properties or even casting.

The cleanest solution is to annotate the queryset to give those keys a name not containing `__`.
After the query has been annotated, we can refer to that name and avoid any collisions.

IA-4652